refactor: mermaidify.tsを4つのファイルに分割して保守性を向上
新規ファイル:
- directoryTreeBuilder.ts: ディレクトリツリー構築ロジック (90行)
- mermaidWriter.ts: Mermaid記法出力処理 (156行)
- mermaidUtils.ts: 文字列変換ユーティリティ (9行)

変更ファイル:
- mermaidify.ts: オーケストレータとして簡潔に (71行)
- mermaidify.fileNameToMermaidId.spec.ts: インポート先をmermaidUtilsに修正

改善効果:
- 単一責任原則の適用により各ファイルが明確な役割を持つ
- 認知的複雑度を20から最大14に削減
- 保守性指数を24.76から最高67.51まで向上
- テスタビリティと再利用性の改善
- 循環依存のないクリーンな階層構造
- 全テスト通過確認済み
